[NFC] Move manifests-related code to `Workspace+Manifests.swift`
`Workspace.swift` is currently at ~4600 LoC and has a risk of growing larger. It has become an assorted collection of stateful code, and splitting it into separate components should make it more manageable.

I started with manifests-related code, moving it `Workspace+Manifests.swift`.
